Move server files to package directory
Move server folders to a new package directory so that both the server starting script and the test runner can find the files.
