+++ server/backend/authentication_manager.py
