+++ server/backend/service_manager.py
@@ -1,5 +1,5 @@
 from typing import MutableMapping
-from server.user_data import UserData
+from backend.user_data import UserData
 
 class ServiceManager:
     """
