+++ server/tests/_request_handler/_add_habit_test.py
@@ -1,10 +1,10 @@
 
 
 import unittest
-from server.authentication_manager import AuthenticationManager
+from backend.authentication_manager import AuthenticationManager
 
-from server.server import _RequestHandler
-from server.service_manager import ServiceManager
+from backend.server import _RequestHandler
+from backend.service_manager import ServiceManager
 
 
 class TestAddHabit(unittest.TestCase):
