+++ server/tests/_request_handler/_get_missing_fields_test.py
@@ -1,7 +1,7 @@
 import unittest
-from server.authentication_manager import AuthenticationManager
-from server.server import _RequestHandler
-from server.service_manager import ServiceManager
+from backend.authentication_manager import AuthenticationManager
+from backend.server import _RequestHandler
+from backend.service_manager import ServiceManager
 
 class TestGetMissingFields(unittest.TestCase):    
     """
