+++ server/tests/_request_handler/handle_request_test.py
@@ -1,8 +1,7 @@
 import unittest
-from urllib import request
-from server.authentication_manager import AuthenticationManager
-from server.server import _RequestHandler
-from server.service_manager import ServiceManager
+from backend.authentication_manager import AuthenticationManager
+from backend.server import _RequestHandler
+from backend.service_manager import ServiceManager
 
 class TestHandleRequest(unittest.TestCase):    
     """
