+++ server/tests/authentication_manager/constructor_test.py
@@ -1,5 +1,5 @@
 import unittest
-from server.authentication_manager import AuthenticationManager
+from backend.authentication_manager import AuthenticationManager
 
 class TestConstructor(unittest.TestCase):    
     """
