+++ server/tests/authentication_manager/get_username_for_token_test.py
@@ -1,5 +1,5 @@
 import unittest
-from server.authentication_manager import AuthenticationManager
+from backend.authentication_manager import AuthenticationManager
 
 class TestGetUsernameForToken(unittest.TestCase):    
     """
