+++ server/tests/habit/constructor_test.py
@@ -1,6 +1,6 @@
 import unittest
 
-from server.habit import Habit
+from backend.habit import Habit
 
 class TestConstructor(unittest.TestCase):    
     """
