+++ server/tests/habit/create_json_dict_test.py
@@ -1,6 +1,6 @@
 import unittest
 
-from server.habit import Habit
+from backend.habit import Habit
 
 class TestCreateJsonDict(unittest.TestCase):    
     """
