+++ server/tests/service_manager/create_user_test.py
@@ -1,5 +1,5 @@
 import unittest
-from server.service_manager import ServiceManager
+from backend.service_manager import ServiceManager
 
 class TestCreateUser(unittest.TestCase):    
     """
