+++ server/tests/service_manager/remove_habit_test.py
@@ -1,6 +1,6 @@
 import unittest
-from server.service_manager import ServiceManager
-from server.user_data import UserData
+from backend.service_manager import ServiceManager
+from backend.user_data import UserData
 
 class TestRemoveHabit(unittest.TestCase):    
     """
