+++ server/tests/user_data/constructor_test.py
@@ -1,5 +1,5 @@
 import unittest
-from server.user_data import UserData
+from backend.user_data import UserData
 
 class TestConstructor(unittest.TestCase):    
     """
