+++ server/tests/user_data/increment_habit_id_test.py
@@ -1,5 +1,5 @@
 import unittest
-from server.user_data import UserData
+from backend.user_data import UserData
 
 class TestIncrementHabitId(unittest.TestCase):    
     """
